pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "remote-embedding"
version = "0.2.0"
description = "A shared FastAPI embedding service and LangChain-compatible remote client for reusing one embedding model across multiple applications and lowering VRAM usage on limited GPUs."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE"]
authors = [
  { name = "Meshkat Shariat Bagheri" }
]
keywords = ["embeddings", "fastapi", "langchain", "huggingface", "api"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Framework :: FastAPI",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "fastapi>=0.115",
  "langchain-core>=0.3",
  "langchain-huggingface>=0.1.2",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "requests>=2.32",
  "uvicorn>=0.30",
]

[project.urls]
Homepage = "https://github.com/MeshkatShB/remote-embedding"
Issues = "https://github.com/MeshkatShB/remote-embedding/issues"

[project.scripts]
remote-embedding-server = "remote_embedding.app:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
